gen jwt, token verif, linking events to users

+++ models/event.go
@@ -6,12 +6,12 @@ import (
 )
 
 type Event struct {
-	ID          int
+	ID          int64
 	Name        string    `binding:"required"`
 	Description string    `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserID      int
+	UserID      int64
 }
 
 func (e *Event) Save() error {
@@ -28,7 +28,7 @@ func (e *Event) Save() error {
 		return err
 	}
 	id, err := result.LastInsertId()
-	e.ID = int(id)
+	e.ID = id
 	return err
 }
